test: continue test refactor
